+++ packages/webview_flutter/webview_flutter_android/lib/src/android_webview.dart
@@ -7,6 +7,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart' show BinaryMessenger;
 
 import 'package:flutter/widgets.dart' show WidgetsFlutterBinding;
+import 'package:webview_flutter_platform_interface/webview_flutter_platform_interface.dart';
 
 import 'android_webview.g.dart';
 import 'android_webview_api_impls.dart';
@@ -939,7 +940,7 @@ class WebViewClient extends JavaObject {
   final void Function(WebView webView, String url, bool isReload)?
       doUpdateVisitedHistory;
 
-  /// This callback is only called for requests that require HTTP basic authentication.
+  /// This callback is only called for requests that require HTTP authentication.
   final void Function(
     WebView webView,
     HttpAuthHandler handler,
@@ -1508,7 +1509,7 @@ class CustomViewCallback extends JavaObject {
   }
 }
 
-/// Represents a request for HTTP basic authentication.
+/// Represents a request for HTTP authentication.
 ///
 /// Instances of this class are created by the [WebView] and passed to
 /// [WebViewClient.onReceivedHttpAuthRequest]. The host application must call
@@ -1532,7 +1533,16 @@ class HttpAuthHandler extends JavaObject {
 
   /// Instructs the WebView to proceed with the authentication with the provided
   /// credentials.
-  Future<void> proceed(String username, String password) {
-    return api.proceedFromInstance(this, username, password);
+  Future<void> proceed(WebViewCredential credential) {
+    return api.proceedFromInstance(this, credential);
+  }
+
+  /// Gets whether the credentials stored for the current host are suitable for
+  /// use.
+  ///
+  /// Credentials are not suitable if they have previously been rejected by the
+  /// server for the current request.
+  Future<bool> useHttpAuthUsernamePassword() {
+    return api.useHttpAuthUsernamePasswordFromInstance(this);
   }
 }