+++ packages/webview_flutter/webview_flutter_platform_interface/test/platform_navigation_delegate_test.dart
@@ -151,8 +151,7 @@ void main() {
             const PlatformNavigationDelegateCreationParams());
 
     expect(
-      () => callbackDelegate
-          .setOnHttpBasicAuthRequest((HttpBasicAuthRequest request) {}),
+      () => callbackDelegate.setOnHttpAuthRequest((HttpAuthRequest request) {}),
       throwsUnimplementedError,
     );
   });